requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.4
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
